requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# ES_Basic_Feeder/Basic_API.py: 1,2
# ES_Basic_Feeder/Basic_Feeder.py: 5,6
elasticsearch == 7.16.2

# ES_Basic_Feeder/DB_Feeder.py: 16
ha-jpype1 == 2.6.6

# ES_Basic_Feeder/DB_Feeder.py: 9
jaydebeapi == 1.2.3

# ES_Basic_Feeder/DB_Feeder.py: 16
jpype1 == 1.3.0

# ES_Basic_Feeder/DB_Feeder.py: 16
jpype1-py3 == 0.5.5.4

# ES_Basic_Feeder/DB_Feeder.py: 10,11,12
pandas == 1.3.4

# ES_Basic_Feeder/Basic_Feeder.py: 3
requests == 2.26.0

# ES_Basic_Feeder/DB_Feeder.py: 19
schedule == 1.1.0
